Реализация команды удаления элемента. Разделение ViewModel для выделения команд.
