+++ DOfficeCore/ToDoList.txt
@@ -1,20 +1,1 @@
-﻿1) Изменить модель:
-	а) Добавить класс, который будет содержать коллекцию диагнозов + 
-	коллекцию текущих кодов диагнозов + коллекцию текущих имён блоков + коллекцию текущих строк
-	б) Добавить методы для изменений содержимого коллекции диагнозов в зависимости от текущих коллекций
-2) Изменить Сервис провайдера данных, чтобы адекватно сохранять новую модель
-
-ИЛИ
-
-1) Добавить класс, который будет содержать только три коллекции со строками
-2) Добавить провайдер для взаимодействия новой модели со старой
-
-ИЛИ
-
-1) Изменить DataGridProvider под соответствующие нужды
-2) ... с помощью LINQ
-
-____________________________________________________
-
-1) Переписать все команды в отдельные файлы
-2) Переписать провайдер сохранения данных и загрузки данных из файла
+﻿Связи во ViewCollection через сеттер
